Moved place/extract_bits into utils.smode.h
